docs: improvements
