+++ @udir-design/react/src/design-tokens/Colors.mdx
@@ -11,19 +11,14 @@ import {
   DatavisSeqColorDisplay,
   DatavisSeqDivColorDisplay,
   DatavisDivGradient,
-} from './docs/color/datavis-color-display';
+} from './docs/color/dataviz-color-display';
 
 <Meta title="design-tokens/Farger" />
+<Unstyled>
 
 # Farger
 
 > Gå til [oversikten over designtokens](/docs/design-tokens-oversikt--docs) for å se CSS-variablene til fargene.
-
-## Farger for andre flater
-
-Dersom du skal lage illustrasjoner, trykksaker eller lignende, kan du lese mer om [fargepaletter for øvrige bruksområder](https://www.udir.no/om-udir/designprofil/farger/) på Udir.no.
-
-Vi deler fargene i designsystemet inn i tre grupper: Farger for brukergrensesnitt, datavisualisering, og for andre flater.
 
 ## Farger for brukergrensesnitt
 
@@ -46,8 +41,6 @@ Identitetsfargene tar utgangspunkt i Udirs visuelle identitet, men nyansene er t
 Varselfargene brukes til å formidle varsler av typen info, suksess, advarsel eller fare. Disse fargene er tilgjengelig for komponenter der det er relevant, for eksempel [`Alert`](/docs/components-alert--docs) og [`Tag`](/docs/components-tag--docs). Varselfarger skal ikke brukes som dekor eller fordi de ser fine ut. De skal kun benyttes i kontekst der innholdet har tilsvarende funksjon.
 
 <SeverityColors />
-
-<Unstyled>
 
 ## Farger for datavisualisering
 
@@ -118,3 +111,9 @@ Numeriske data med avvik fra et midtpunkt bruker en divergerende palett med to m
 **Linjediagram:** Bruk geometriske figurer (sirkler, trekanter, firkanter) knyttet til linjene i tillegg til farge for å gjøre det enklere å oppfatte forskjeller.
 
 </Unstyled>
+
+## Farger for andre flater
+
+Dersom du skal lage illustrasjoner, trykksaker eller lignende, kan du lese mer om [fargepaletter for øvrige bruksområder](https://www.udir.no/om-udir/designprofil/farger/) på Udir.no.
+
+Vi deler fargene i designsystemet inn i tre grupper: Farger for brukergrensesnitt, datavisualisering, og for andre flater.
