+++ @udir-design/react/src/utilities/dataVisualization/highcharts.ts
@@ -28,23 +28,28 @@ import { getCategoricalColors } from './dataVisualization';
  * import HighchartsReact from 'highcharts-react-official';
  * import { getHighchartsTheme } from '@udir-design/react/utilities/dataVisualization/alpha';
  * import '@udir-design/theme/dataviz.css';
- * import 'highcharts/i18n/nb-NO'; // Optional: for Norwegian
+ * import 'highcharts/i18n/nb-NO'; // Norwegian Bokmål language pack
  *
- * const theme = getHighchartsTheme();
+ * Highcharts.setOptions(getHighchartsTheme());
  *
- * const chartOptions: Highcharts.Options = {
- *   ...theme,
- *   chart: {
- *     ...theme.chart,
- *     type: 'pie',
- *   },
- *   series: [{
- *     type: 'pie',
- *     data: [1, 2, 3],
- *   }],
- * };
+ * export function MyChart() {
+ *   const options: Highcharts.Options = {
+ *     chart: { type: 'pie' },
+ *     title: { text: 'Prøveresultater' },
+ *     series: [
+ *       {
+ *         type: 'pie',
+ *         name: 'Antall',
+ *         data: [
+ *           { name: 'Bestått', y: 300 },
+ *           { name: 'Ikke bestått', y: 50 },
+ *         ],
+ *       },
+ *     ],
+ *   };
  *
- * <HighchartsReact highcharts={Highcharts} options={chartOptions} />
+ *   return <HighchartsReact highcharts={Highcharts} options={options} />;
+ * }
  * ```
  */
 export function getHighchartsTheme(): Partial<Highcharts.Options> {
